refactor: extract chrome-tools.ts (step 1/N of #274 split)
First step in restructuring the tool layer per owner's request.
Re-does PR #274's intent on top of current main (PR #274 was
massively stale and had scope creep).

Moved to new src/chrome-tools.ts:
- switchTabTool (was in browser-tools.ts)
- closeTabTool (was in browser-tools.ts, from PR #300)
- openUrlTool (was in inline-tools.ts)
- TAB_ALIASES helper

browser-tools.ts re-exports for backward compat, so all existing
imports keep working without changes.

This is intentionally a small first step. If the structure is
agreeable, follow-up PRs will extract recording-tools.ts and
remote-meeting-tools.ts in separate commits.
